fix gh install dialog showing undefined initially
